PR qui résume tout ce que j'ai implémenté et/ou modifié dans les
fichiers pour la gestion et l'affichage des **catégories**.

### CategorieController.php

Changement des routes pour certaines méthodes ( *index, new, edit,
delete* ) pour un affichage dans le back-office (admin/)

- `index` : Affiche toutes les catégories.
- `new` : Crée une nouvelle catégorie. Génération du `slug` à partir du
nom donné.
- `edit` :  modification d'une catégorie : 
- si le formulaire est soumis et valide, le slug est mis à jour si le
nom a changé.
  - Création d'une nouvelle instance pour la propriété `updatedAt`. 
- `delete` : Supprime la catégorie liée à l'ID. 

Ces quatre méthodes sont accessibles uniquement aux utilisateurs ayant
au moins `ROLE_ADMIN`.
Si l'utilisateur n'a que le ROLE_USER ou n'est pas connecté, il est
redirigé à la page d'accueil.

- `show` : affichage des catégorie ( mappé par le slug ).  
- Affichage granulaire en fonction des droits d'utilisateur. Les
utilisateurs avec le `ROLE_USER` et les utilisateurs non connectés
voient uniquement le `nom` et la `description` des catégories.

### Categorie.php 
A présent que le slug est configuré et fonctionnel, il ne peut plus être
nullable
- `nullable = true` a été effacé.

- Création d'un fichier de migration pour mettre à jour le schéma de la
table `Catégorie` pour la propriété `slug`.

C'est pour cette raison que la propriété `nom` est unique, pour garantir
l'unicité de chaque enregistrement présent dans `slug`.

### CategorieType.php

- Implémentation détaillée du formulaire liée à l'entité `Categorie`
- Création des messages à afficher si le champs obligatoire est vide ou
si les validations ne sont pas correctes.
- Mise en place du token *CSRF* 

#### Les champs qui s'affichent lors de la création ou bien lors de la
modification d'une *catégorie* :
- `nom` : champ requis et unique 
- `description` : optionnel

## Templates 
### templates/_partials/_navbar 
- remplacer `categorie.id` -> `categorie.slug`

### templates/categorie/index.html.twig
- remplacer les boutons/liens de navigation :  
`categorie.id` => `categorie.slug`

### templates/categorie/new.html.twig 
- changement du titre

### temapltes/categorie/show.html.twig
- changement des conditions pour l'affichage en fonction des droits
utilisateurs.

### app.css
- ajout du style pour les tag `textarea`

+++ src/Controller/CategorieController.php
@@ -5,31 +5,43 @@ namespace App\Controller;
 use App\Entity\Categorie;
 use App\Form\CategorieType;
 use App\Repository\CategorieRepository;
+use DateTimeImmutable;
 use Doctrine\ORM\EntityManagerInterface;
+use Symfony\Bridge\Doctrine\Attribute\MapEntity;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Attribute\Route;
+use Symfony\Component\String\Slugger\SluggerInterface;
 
 // #[Route('/categorie')]
 final class CategorieController extends AbstractController
 {
-    #[Route('/categorie', name: 'app_categorie_index', methods: ['GET'])]
+    #[Route('admin/categorie', name: 'app_categorie_index', methods: ['GET'])]
     public function index(CategorieRepository $categorieRepository): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) { // Redirige vers la route de la page d'accueil 
+            return $this->redirectToRoute('app_main');
+        }
         return $this->render('categorie/index.html.twig', [
             'categories' => $categorieRepository->findAll(),
         ]);
     }
 
     #[Route('/admin/categorie/new', name: 'app_categorie_new', methods: ['GET', 'POST'])]
-    public function new(Request $request, EntityManagerInterface $entityManager): Response
+    public function new(Request $request, EntityManagerInterface $entityManager,  SluggerInterface $slugger): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) { // Redirige vers la route de la page d'accueil 
+            return $this->redirectToRoute('app_main');
+        }
         $categorie = new Categorie();
         $form = $this->createForm(CategorieType::class, $categorie);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            $slugCategorie = $slugger->slug($categorie->getNom());
+            // dump($slugCategorie);
+            $categorie->setSlug(strtolower($slugCategorie));
             $entityManager->persist($categorie);
             $entityManager->flush();
 
@@ -42,22 +54,35 @@ final class CategorieController extends AbstractController
         ]);
     }
 
-    #[Route('/categorie/{id}', name: 'app_categorie_show', methods: ['GET'])]
-    public function show(Categorie $categorie): Response
+    // #[Route('/categorie/{id}', name: 'app_categorie_show', methods: ['GET'])]
+    #[Route('/categorie/{slug}', name: 'app_categorie_show', methods: ['GET'])]
+    public function show(#[MapEntity(mapping: ['slug' => "slug"])] Categorie $categorie, SluggerInterface $slugger): Response
     {
+        // $slugCategorie = $slugger->slug($categorie->getNom());
+        // $categorie->setSlug(strtolower($slugCategorie));
+        // dump($slugCategorie);
+        // dd($categorie);
 
         return $this->render('categorie/show.html.twig', [
             'categorie' => $categorie,
         ]);
     }
 
-    #[Route('/admin/categorie/{id}/edit', name: 'app_categorie_edit', methods: ['GET', 'POST'])]
-    public function edit(Request $request, Categorie $categorie, EntityManagerInterface $entityManager): Response
+    #[Route('/admin/categorie/{slug}/edit', name: 'app_categorie_edit', methods: ['GET', 'POST'])]
+    public function edit(Request $request, #[MapEntity(mapping: ['slug' => "slug"])] Categorie $categorie, EntityManagerInterface $entityManager, SluggerInterface $slugger): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) { // Redirige vers la route de la page d'accueil 
+            return $this->redirectToRoute('app_main');
+        }
         $form = $this->createForm(CategorieType::class, $categorie);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            $nomToSlug = $form->get('nom')->getData();
+            $slugNom = $slugger->slug($nomToSlug);
+            $categorie->setSlug(strtolower($slugNom));
+            $categorie->setUpdatedAt(new DateTimeImmutable());
+            $entityManager->persist($categorie);
             $entityManager->flush();
 
             return $this->redirectToRoute('app_categorie_index', [], Response::HTTP_SEE_OTHER);
@@ -72,6 +97,9 @@ final class CategorieController extends AbstractController
     #[Route('/admin/categorie/{id}', name: 'app_categorie_delete', methods: ['POST'])]
     public function delete(Request $request, Categorie $categorie, EntityManagerInterface $entityManager): Response
     {
+        if (!$this->isGranted('ROLE_ADMIN')) { // Redirige vers la route de la page d'accueil 
+            return $this->redirectToRoute('app_main');
+        }
         if ($this->isCsrfTokenValid('delete' . $categorie->getId(), $request->getPayload()->getString('_token'))) {
             $entityManager->remove($categorie);
             $entityManager->flush();
